+++ backend/src/modules/addresses/index.ts
@@ -1,1 +1,4 @@
-export {};
+export * from "./address.dto";
+export * from "./address.validation";
+export * from "./address.controller";
+export { default as addressesRouter } from "./address.routes";